Add feature "derive" to reexport the macros
This enables these derives to be qualified under `npy`, for
disambiguation from `serde`:

    extern crate npy;

    #[derive(npy::Serialize, npy::Deserialize)]
    struct MyStruct { ... }

This has a couple of downsides with regard to maintainence:

* npy_derive can no longer be a dev-dependency because it must
  become an optional dependency.
* Many tests and examples need the feature.  We need to list all
  of these in Cargo.toml.
* Because this crate is 2015 edition, as soon as we list *any*
  tests and examples, we must list *all* of them; including the
  ones that don't need the feature!

---

This commit had to update `.travis.yml` to start using the feature.
I took this opportunity to also add `--examples` (which the default
script does not do) to ensure that examples build correctly.