add a benchmark for to_vec
I tried a variety of things to optimize this function:

* Replacing usage of get_unchecked with reuse of the remainder returned
  by read_one, so that the stride can be statically known rather than
  having to be looked up. (this is what optimized the old read benchmark)
* Putting an assertion up front to prove that the data vector is long enough.

But whatever I do, performance won't budge.  In the f32 benchmark, a very
hot bounds check still occurs on every read to ensure that the length of
the data is at least 4 bytes.

So I'm adding the benchmark, but leaving the function itself alone.