remove size1_cfg
It was an artefact of an old design.

Out of paranoia, I added some assertions to the Serialize/Deserialize
impls to make sure the endianness is valid.  These are redundant since
it is checked in TypeStr::from_str, but most other such properties are
at least implicitly checked by the `_` arm in impls of `reader` and
`writer` and I wanted to be safe.